Move some code from assign_from_json to new file
This is in preparation for adding a new class in this file, we add there
the function and class which will be needed by this class.
