+++ jbmc/src/java_bytecode/assignments_from_json.cpp
@@ -9,6 +9,7 @@ Author: Diffblue Ltd.
 #include "assignments_from_json.h"
 
 #include "ci_lazy_methods_needed.h"
+#include "code_with_references.h"
 #include "java_static_initializers.h"
 #include "java_string_library_preprocess.h"
 #include "java_string_literals.h"
@@ -418,22 +419,6 @@ static void assign_array_data_component_from_json(
       array_init_data, from_integer(index, java_int_type()));
     assign_from_json_rec(element_at_index, *it, inferred_element_type, info);
   }
-}
-
-/// Allocate a fresh array of length \p array_length_expr and assigns \p expr
-/// to it.
-static codet allocate_array(
-  const exprt &expr,
-  const exprt &array_length_expr,
-  const source_locationt &loc)
-{
-  const pointer_typet &pointer_type = to_pointer_type(expr.type());
-  const auto &element_type =
-    java_array_element_type(to_struct_tag_type(pointer_type.subtype()));
-  side_effect_exprt java_new_array{ID_java_new_array, pointer_type};
-  java_new_array.copy_to_operands(array_length_expr);
-  java_new_array.type().subtype().set(ID_element_type, element_type);
-  return code_assignt{expr, java_new_array, loc};
 }
 
 /// Declare a non-deterministic length expression
